Move seismic services into src/models/stores/seismic/.
