+++ src/models/stores/seismic/seismic-download-worker-runner.ts
@@ -2,12 +2,12 @@
 // CommonJS loader cannot parse) never ends up in a module Jest loads eagerly. The
 // service reaches this only through a dynamic import from its default runner, and tests
 // inject their own runner, so Jest never loads this file.
-import { DownloadEvent, DownloadRunner } from "../../../shared/seismic/seismic-downloader";
-import { getLocalBaseUrl, isProxyEnabled } from "../../../shared/seismic/earthscope-client";
+import { DownloadEvent, DownloadRunner } from "../../../../shared/seismic/seismic-downloader";
+import { getLocalBaseUrl, isProxyEnabled } from "../../../../shared/seismic/earthscope-client";
 
 /** Default runner: spawns the download Web Worker and forwards its events. */
 export const workerRunner: DownloadRunner = (params, onEvent, cancel) => {
-  const worker = new Worker(new URL("../../workers/seismic-download-worker.ts", import.meta.url));
+  const worker = new Worker(new URL("../../../workers/seismic-download-worker.ts", import.meta.url));
 
   worker.onmessage = (e: MessageEvent<DownloadEvent>) => {
     onEvent(e.data);
