+++ src/models/stores/seismic/seismic-query-service.test.ts
@@ -1,9 +1,9 @@
 import { DateTime } from "luxon";
 import { SeismicQueryService, envelopeCacheKey } from "./seismic-query-service";
-import { encodeEnvelopeTile } from "../../../shared/seismic/envelope-codec";
-import { LEVEL_SPACINGS, NO_DATA_SENTINEL } from "../../../shared/seismic/envelope-config";
-import { getTileTimeRange } from "../../../shared/seismic/tile-addressing";
-import { SeismicViewportParams } from "../../../shared/seismic/seismic-types";
+import { encodeEnvelopeTile } from "../../../../shared/seismic/envelope-codec";
+import { LEVEL_SPACINGS, NO_DATA_SENTINEL } from "../../../../shared/seismic/envelope-config";
+import { getTileTimeRange } from "../../../../shared/seismic/tile-addressing";
+import { SeismicViewportParams } from "../../../../shared/seismic/seismic-types";
 
 // Mock fetch globally
 const mockFetch = jest.fn();
